Revert "사라진 파일들 다시 받아서 백업"
This reverts commit b5911de7a89c815f293dd97005e07b7172f8a662.

+++ moneytree_back/src/main/java/com/moneytree_back/domain/DepositAccount.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -30,7 +29,7 @@ public class DepositAccount {
     // @Column(name = "dep") // 멤버는 보류
 
     @Column(name = "deposit_amount", nullable = false, precision = 15, scale = 2)
-    private BigDecimal depositAmount;
+    private Double depositAmount;
 
     @Column(name = "deposit_start_date", nullable = false)
     private LocalDate depositStartDate;
