+++ moneytree_back/src/main/java/com/moneytree_back/domain/DepositRate.java
@@ -5,7 +5,6 @@ import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -27,10 +26,10 @@ public class DepositRate {
     private DepositProduct depositProduct; // 외래 키
 
     @Column(name = "deposit_interest_rate", nullable = false, precision = 5, scale = 2)
-    private BigDecimal depositInterestRate;
+    private Double depositInterestRate;
 
     @Column(name = "deposit_special_rate", precision = 5, scale = 2)
-    private BigDecimal depositSpecialRate; // 사용 여부는 선택
+    private Double depositSpecialRate; // 사용 여부는 선택
 
     @CreationTimestamp
     @Column(name = "deposit_rate_created_at", updatable = false)
